Split off factory module
