+++ packages/liveblocks-react/src/index.tsx
@@ -18,7 +18,7 @@ export {
   useRedo,
   useBatch,
   useHistory,
-} from "./rooms";
+} from "./factory";
 
 // Re-exports from @liveblocks/client, for convenience
 export type { Json, JsonObject } from "@liveblocks/client";
